refactor(cli): split the setup god-module into composed units
cli/setup.py was a 647-LOC god-module mixing six unrelated concerns: main-clone resolution, global t3 install, APM dependency install + apm-hook stripping, runtime skill-symlink syncing, Claude-plugin registration, and the command coordination itself.

Split into a cli/setup/ package, one concern per module: clone.py (find_main_clone / validate_repo), tool_installer.py (ToolInstaller), apm.py (ApmInstaller + strip_apm_hooks), skill_linker.py (SkillLinker), plugin_registrar.py (PluginRegistrar), _process.py (shared run_captured helper), command.py (the thin t3 setup typer command, coordination only), __init__.py (re-exports only).

Behavior-preserving: logic moved verbatim into the new units; t3 setup installs / registers / links exactly what it did before. Stateless helpers are @staticmethod; the command stays typer-runner typer.Exit (not a call_command path). All call sites updated (update.py, every test patch target) with no stale references. Module-health is now green: every new module is well under the 500-LOC cap.
