+++ tests/cli_setup/test_skill_symlinks.py
@@ -10,13 +10,7 @@ from pathlib import Path
 from types import SimpleNamespace
 from unittest.mock import patch
 
-from teatree.cli.setup import (
-    CORE_EXCLUDED_SKILLS,
-    _clean_broken_symlinks,
-    _ensure_skill_link,
-    _remove_excluded_skills,
-    _sync_skill_symlinks,
-)
+from teatree.cli.setup.skill_linker import CORE_EXCLUDED_SKILLS, SkillLinker, _ensure_skill_link
 
 
 class TestRemoveExcludedSkills:
@@ -25,31 +19,31 @@ class TestRemoveExcludedSkills:
         target.mkdir()
         link = tmp_path / "using-superpowers"
         link.symlink_to(target)
-        assert _remove_excluded_skills(tmp_path, ["using-superpowers"]) == 1
+        assert SkillLinker(tmp_path, tmp_path).remove_excluded(["using-superpowers"]) == 1
         assert not link.exists()
 
     def test_removes_directories(self, tmp_path: Path) -> None:
         skill_dir = tmp_path / "using-git-worktrees"
         skill_dir.mkdir()
         (skill_dir / "SKILL.md").touch()
-        assert _remove_excluded_skills(tmp_path, ["using-git-worktrees"]) == 1
+        assert SkillLinker(tmp_path, tmp_path).remove_excluded(["using-git-worktrees"]) == 1
         assert not skill_dir.exists()
 
     def test_ignores_nonexistent(self, tmp_path: Path) -> None:
-        assert _remove_excluded_skills(tmp_path, ["nonexistent"]) == 0
+        assert SkillLinker(tmp_path, tmp_path).remove_excluded(["nonexistent"]) == 0
 
     def test_multiple_excluded(self, tmp_path: Path) -> None:
         target = tmp_path / "target"
         target.mkdir()
         (tmp_path / "skill-a").symlink_to(target)
         (tmp_path / "skill-b").mkdir()
-        assert _remove_excluded_skills(tmp_path, ["skill-a", "skill-b", "skill-c"]) == 2
+        assert SkillLinker(tmp_path, tmp_path).remove_excluded(["skill-a", "skill-b", "skill-c"]) == 2
 
     def test_rejects_path_traversal(self, tmp_path: Path) -> None:
-        assert _remove_excluded_skills(tmp_path, ["../etc", ".hidden"]) == 0
+        assert SkillLinker(tmp_path, tmp_path).remove_excluded(["../etc", ".hidden"]) == 0
 
     def test_rejects_nested_path(self, tmp_path: Path) -> None:
-        assert _remove_excluded_skills(tmp_path, ["foo/bar"]) == 0
+        assert SkillLinker(tmp_path, tmp_path).remove_excluded(["foo/bar"]) == 0
 
 
 class TestEnsureSkillLink:
@@ -153,10 +147,10 @@ class TestSyncSkillSymlinks:
 
         with (
             patch("teatree.agents.skill_bundle.DEFAULT_SKILLS_DIR", skills_src),
-            patch("teatree.cli.setup.DoctorService") as mock_svc,
+            patch("teatree.cli.setup.skill_linker.DoctorService") as mock_svc,
         ):
             mock_svc.collect_overlay_skills.return_value = []
-            created, fixed = _sync_skill_symlinks(claude_skills, tmp_path / "workspace")
+            created, fixed = SkillLinker(claude_skills, tmp_path / "workspace").sync()
 
         assert created == 2
         assert fixed == 0
@@ -176,10 +170,10 @@ class TestSyncSkillSymlinks:
 
         with (
             patch("teatree.agents.skill_bundle.DEFAULT_SKILLS_DIR", skills_src),
-            patch("teatree.cli.setup.DoctorService") as mock_svc,
+            patch("teatree.cli.setup.skill_linker.DoctorService") as mock_svc,
         ):
             mock_svc.collect_overlay_skills.return_value = [(overlay_skill, "my-skill")]
-            created, _fixed = _sync_skill_symlinks(claude_skills, tmp_path / "workspace")
+            created, _fixed = SkillLinker(claude_skills, tmp_path / "workspace").sync()
 
         assert created == 1
         assert (claude_skills / "my-skill").is_symlink()
@@ -195,10 +189,10 @@ class TestSyncSkillSymlinks:
 
         with (
             patch("teatree.agents.skill_bundle.DEFAULT_SKILLS_DIR", skills_src),
-            patch("teatree.cli.setup.DoctorService") as mock_svc,
+            patch("teatree.cli.setup.skill_linker.DoctorService") as mock_svc,
         ):
             mock_svc.collect_overlay_skills.return_value = []
-            created, fixed = _sync_skill_symlinks(runtime_skills, tmp_path / "workspace", sync_core=False)
+            created, fixed = SkillLinker(runtime_skills, tmp_path / "workspace").sync(sync_core=False)
 
         assert created == 0
         assert fixed == 0
@@ -216,10 +210,10 @@ class TestSyncSkillSymlinks:
 
         with (
             patch("teatree.agents.skill_bundle.DEFAULT_SKILLS_DIR", skills_src),
-            patch("teatree.cli.setup.DoctorService") as mock_svc,
+            patch("teatree.cli.setup.skill_linker.DoctorService") as mock_svc,
         ):
             mock_svc.collect_overlay_skills.return_value = []
-            _sync_skill_symlinks(runtime_skills, tmp_path / "workspace", sync_core=False)
+            SkillLinker(runtime_skills, tmp_path / "workspace").sync(sync_core=False)
 
         assert not (runtime_skills / "code").exists()
 
@@ -237,10 +231,10 @@ class TestSyncSkillSymlinks:
 
         with (
             patch("teatree.agents.skill_bundle.DEFAULT_SKILLS_DIR", skills_src),
-            patch("teatree.cli.setup.DoctorService") as mock_svc,
+            patch("teatree.cli.setup.skill_linker.DoctorService") as mock_svc,
         ):
             mock_svc.collect_overlay_skills.return_value = [(overlay_skill, "my-skill")]
-            created, _fixed = _sync_skill_symlinks(runtime_skills, tmp_path / "workspace", sync_core=False)
+            created, _fixed = SkillLinker(runtime_skills, tmp_path / "workspace").sync(sync_core=False)
 
         assert created == 1
         assert (runtime_skills / "my-skill").is_symlink()
@@ -270,7 +264,7 @@ class TestSetupSyncsCodexWhenDirExists:
         Setup should symlink core skills into ~/.codex/skills but NOT into
         ~/.claude/skills.
         """
-        from teatree.cli import setup as setup_module  # noqa: PLC0415
+        from teatree.cli.setup import command as setup_module  # noqa: PLC0415
 
         skills_src = tmp_path / "core_skills"
         skills_src.mkdir()
@@ -292,11 +286,11 @@ class TestSetupSyncsCodexWhenDirExists:
 
         with (
             patch("teatree.agents.skill_bundle.DEFAULT_SKILLS_DIR", skills_src),
-            patch.object(setup_module, "_find_main_clone", return_value=repo),
-            patch.object(setup_module, "_run_apm_install", return_value=True),
-            patch.object(setup_module, "_install_claude_plugin", return_value=True),
+            patch.object(setup_module, "find_main_clone", return_value=repo),
+            patch.object(setup_module, "ApmInstaller"),
+            patch.object(setup_module, "PluginRegistrar"),
             patch.object(setup_module, "ensure_self_db_migrated", return_value=False),
-            patch.object(setup_module, "DoctorService") as mock_svc,
+            patch("teatree.cli.setup.skill_linker.DoctorService") as mock_svc,
             patch("teatree.config.load_config") as mock_load,
         ):
             mock_svc.collect_overlay_skills.return_value = []
@@ -315,7 +309,7 @@ class TestSetupSyncsCodexWhenDirExists:
         teatree versions; they must be pruned so they don't shadow the
         plugin's copies of the same skills.
         """
-        from teatree.cli import setup as setup_module  # noqa: PLC0415
+        from teatree.cli.setup import command as setup_module  # noqa: PLC0415
 
         skills_src = tmp_path / "core_skills"
         skills_src.mkdir()
@@ -336,11 +330,11 @@ class TestSetupSyncsCodexWhenDirExists:
 
         with (
             patch("teatree.agents.skill_bundle.DEFAULT_SKILLS_DIR", skills_src),
-            patch.object(setup_module, "_find_main_clone", return_value=repo),
-            patch.object(setup_module, "_run_apm_install", return_value=True),
-            patch.object(setup_module, "_install_claude_plugin", return_value=True),
+            patch.object(setup_module, "find_main_clone", return_value=repo),
+            patch.object(setup_module, "ApmInstaller"),
+            patch.object(setup_module, "PluginRegistrar"),
             patch.object(setup_module, "ensure_self_db_migrated", return_value=False),
-            patch.object(setup_module, "DoctorService") as mock_svc,
+            patch("teatree.cli.setup.skill_linker.DoctorService") as mock_svc,
             patch("teatree.config.load_config") as mock_load,
         ):
             mock_svc.collect_overlay_skills.return_value = []
@@ -353,7 +347,7 @@ class TestSetupSyncsCodexWhenDirExists:
 
     def test_skips_codex_when_dir_missing(self, tmp_path: Path, monkeypatch) -> None:
         """Setup does not create ~/.codex/skills if it doesn't already exist."""
-        from teatree.cli import setup as setup_module  # noqa: PLC0415
+        from teatree.cli.setup import command as setup_module  # noqa: PLC0415
 
         skills_src = tmp_path / "core_skills"
         skills_src.mkdir()
@@ -373,11 +367,11 @@ class TestSetupSyncsCodexWhenDirExists:
 
         with (
             patch("teatree.agents.skill_bundle.DEFAULT_SKILLS_DIR", skills_src),
-            patch.object(setup_module, "_find_main_clone", return_value=repo),
-            patch.object(setup_module, "_run_apm_install", return_value=True),
-            patch.object(setup_module, "_install_claude_plugin", return_value=True),
+            patch.object(setup_module, "find_main_clone", return_value=repo),
+            patch.object(setup_module, "ApmInstaller"),
+            patch.object(setup_module, "PluginRegistrar"),
             patch.object(setup_module, "ensure_self_db_migrated", return_value=False),
-            patch.object(setup_module, "DoctorService") as mock_svc,
+            patch("teatree.cli.setup.skill_linker.DoctorService") as mock_svc,
             patch("teatree.config.load_config") as mock_load,
         ):
             mock_svc.collect_overlay_skills.return_value = []
@@ -406,14 +400,14 @@ class TestPrunesStaleSkillLinks:
 
         with (
             patch("teatree.agents.skill_bundle.DEFAULT_SKILLS_DIR", skills_src),
-            patch("teatree.cli.setup.DoctorService") as mock_svc,
+            patch("teatree.cli.setup.skill_linker.DoctorService") as mock_svc,
         ):
             mock_svc.collect_overlay_skills.return_value = []
-            _sync_skill_symlinks(runtime, tmp_path / "workspace")
+            SkillLinker(runtime, tmp_path / "workspace").sync()
             removed = self._make_core_skill(skills_src, "full-speed")
-            _sync_skill_symlinks(runtime, tmp_path / "workspace")
+            SkillLinker(runtime, tmp_path / "workspace").sync()
             shutil.rmtree(removed)
-            _sync_skill_symlinks(runtime, tmp_path / "workspace")
+            SkillLinker(runtime, tmp_path / "workspace").sync()
 
         assert {p.name for p in runtime.iterdir()} == {"speed"}
 
@@ -426,12 +420,12 @@ class TestPrunesStaleSkillLinks:
 
         with (
             patch("teatree.agents.skill_bundle.DEFAULT_SKILLS_DIR", skills_src),
-            patch("teatree.cli.setup.DoctorService") as mock_svc,
+            patch("teatree.cli.setup.skill_linker.DoctorService") as mock_svc,
         ):
             mock_svc.collect_overlay_skills.return_value = []
-            _sync_skill_symlinks(runtime, tmp_path / "workspace")
+            SkillLinker(runtime, tmp_path / "workspace").sync()
             (skills_src / "full-speed").rename(skills_src / "speed")
-            _sync_skill_symlinks(runtime, tmp_path / "workspace")
+            SkillLinker(runtime, tmp_path / "workspace").sync()
 
         assert {p.name for p in runtime.iterdir()} == {"speed"}
 
@@ -447,14 +441,14 @@ class TestPrunesStaleSkillLinks:
 
         with (
             patch("teatree.agents.skill_bundle.DEFAULT_SKILLS_DIR", skills_src),
-            patch("teatree.cli.setup.DoctorService") as mock_svc,
+            patch("teatree.cli.setup.skill_linker.DoctorService") as mock_svc,
         ):
             mock_svc.collect_overlay_skills.return_value = [(overlay_skill, "my-skill")]
-            _sync_skill_symlinks(runtime, tmp_path / "workspace")
+            SkillLinker(runtime, tmp_path / "workspace").sync()
             assert (runtime / "my-skill").is_symlink()
             shutil.rmtree(overlay_skill)
             mock_svc.collect_overlay_skills.return_value = []
-            _sync_skill_symlinks(runtime, tmp_path / "workspace")
+            SkillLinker(runtime, tmp_path / "workspace").sync()
 
         assert not (runtime / "my-skill").exists()
 
@@ -470,10 +464,10 @@ class TestPrunesStaleSkillLinks:
 
         with (
             patch("teatree.agents.skill_bundle.DEFAULT_SKILLS_DIR", skills_src),
-            patch("teatree.cli.setup.DoctorService") as mock_svc,
+            patch("teatree.cli.setup.skill_linker.DoctorService") as mock_svc,
         ):
             mock_svc.collect_overlay_skills.return_value = []
-            _sync_skill_symlinks(runtime, tmp_path / "workspace")
+            SkillLinker(runtime, tmp_path / "workspace").sync()
 
         assert (runtime / "my-own-skill").is_dir()
         assert not (runtime / "my-own-skill").is_symlink()
@@ -492,10 +486,10 @@ class TestPrunesStaleSkillLinks:
 
         with (
             patch("teatree.agents.skill_bundle.DEFAULT_SKILLS_DIR", skills_src),
-            patch("teatree.cli.setup.DoctorService") as mock_svc,
+            patch("teatree.cli.setup.skill_linker.DoctorService") as mock_svc,
         ):
             mock_svc.collect_overlay_skills.return_value = []
-            _sync_skill_symlinks(runtime, workspace)
+            SkillLinker(runtime, workspace).sync()
 
         assert (runtime / "code").resolve() == contrib_target.resolve()
 
@@ -503,13 +497,13 @@ class TestPrunesStaleSkillLinks:
 class TestCleanBrokenSymlinks:
     def test_removes_broken(self, tmp_path: Path) -> None:
         (tmp_path / "broken").symlink_to(tmp_path / "nonexistent")
-        assert _clean_broken_symlinks(tmp_path) == 1
+        assert SkillLinker(tmp_path, tmp_path).clean_broken() == 1
 
     def test_leaves_valid(self, tmp_path: Path) -> None:
         target = tmp_path / "target"
         target.mkdir()
         (tmp_path / "valid").symlink_to(target)
-        assert _clean_broken_symlinks(tmp_path) == 0
+        assert SkillLinker(tmp_path, tmp_path).clean_broken() == 0
 
 
 class TestCoreExcludedSkills:
